Fix #1
This commit fixes the issues when sendMaessage was
added with parse_mode. The parse_mode option was
enabled and was causing error while parsing mail.
Due to which telegram API was throwing parsing error
and was responding with BAD_REQUEST.

This commit simply  removes that optional
parse_mode and uses plain text data to send incoming
mail messages.